Use fallible allocation APIs
For the implementation of `Write` for `Vec<u8>`, use `Vec::try_reserve`
to try and allocate memory before copying a slice into the `Vec`. If a
`TryReserveError` is returned, bubble that up to the caller instead of
letting `alloc` do a process abort.

If an allocation fails, return `Error::OutOfMemory`.

This commit removes `derive(Copy)` from the error type in this crate
because the inner `TryReserveError` in the `OutOfMemory` variant is not
`Copy`.