Provide a core::fmt::Write strftime implementation
This implementation uses the same approach as in #49 for
`std::io::Write` and applies it to `core::fmt::Write`. An adapter type
that takes an inner `&mut dyn core::fmt::Write` has
`crate::format::Write` implemented for it.

Much like the implementation of `strftime::string::strftime` which
asserts UTF-8 well-formedness with a call to `expect`, the
implementation of `crate::format::Write` for `FmtWrite` calls
`str::from_utf8(..).expect(..)` on the byte slice passed to it. This
UTF-8 slice is then passed to `core::fmt::Write::write_str`.

This call to expect is not expected to panic for the same reasons why
the call to expect is safe for `strftime::string::strftime`.